Move footnote links to text, changes colours
There wasn't really any point putting just a link in the footnote, and my system of using ::after to add the text to this link was causing accessibility problems. This solutions just makes more sense.

Heading background is darker to meet minimum accessibility requirements (AA). Link text in articles is now slightly darker/lighter (light/dark mode) when not hovered to differentiate it.