Localise URLs, remove testPage from navbar
URLs are now localised rather than always being in English, and language switcher correctly switches between these URLs use the new localiseURL filter. If an article does not exist in one language it will instead redirect to blog page.

testPage has been removed from navbar as it's currently unfinished and is not localised so would not exist when trying to load in another language.