Move article texts into parent folder
This allows images and other article-related files to be stored in the same folder. I think I'll store these in the English folders to keep them in one place, but for consistency I've used the same folder structure with the Spanish articles.

This doesn't affect page URLs, nor language switching