Add viewport cover metatag and padding
This means that my site will be drawn behind the navigation bar on Android (when using gesture navigation instead of 3 button navigation) and on iOS, which I think looks better and feels more immersive (especially in landscape). I believe it also means that the site will be drawn behind the notch on iOS.

All content is shifted to be above the safe-area-(max-)inset-bottom using bottom-padding on the HTML element. This means that the bottom of my site will not be covered by the navigation bar when scrolled all the way down, but will effectively have the same padding as before. The body also has left and right padding to avoid behind drawn behind the notch on iOS. This is done in body so that the header can be, which I think should look fine.

Drawing behind the navigation bar is supported on Chromium (upcoming stable release - default when scrolling down, opt-in for permanent use) and Safari (always since 2017). Chromium supports the max variant of safe-area-inset-bottom, which is generally better to use as it's guaranteed to not change and so cause layout thrashing. Safari and Firefox do not, but also do not recalculate these values.