Rewrite navbar code to flexbox
This avoids using lots of floats and strange HTML ordering. Now the element are placed in the correct order with the default tabindex, and are spaced out using the flexbox (navmenu at 100% shrunk down to the available space after the other elements).

I've removed all the old css code for styling the navbar which is no longer used. I forgot to change the height of the language switcher checkbox when I made the navbar bigger, but it doesn't need to be set now as it just inherits it.