Move wallace to bottom of page for Spanish blog
Had to use position fixed for this because the page stops before the end of the screen as it's not got so much content. But equally doing this on the English page means it's at the bottom right of the viewport as opposed to at the bottom of the page after scrolling down. It's not a perfect solution but more or less works. When I've added more posts to the Spanish one I'll go back to just placing at the end of the page.