Add contents+heading ID plugins
Custom style had to be moved to separate tag because otherwise gets included in contents. I've renamed the old style tag to stylesheet to differentiate them. One is for loading an external file, other is for loading inline CSS.

There's now some 11ty filters for splitting the article into introduction (pre-contents section) and main content, which was the best way I found of doing this. In order to do so, I use the text "%content%" in the article to split it, and the old "contents" tag is no longer used.

The contents is generated by the eleventy-plugin-toc, using mardown-it/-anchor for heading IDs.

Some minor tweaks to the transport article as well. Got rid of the first title as I think it introduces itself, looks neater without. Typo statics --> statistics in the first line. Add aspect-ratio to first image to avoid layout shift.

I've also made a couple formatting tweaks to the entire site. Now the date in the article will be a tad further away from the title, they sometimes felt too cramped before. And now all underlines will have slightly more space between them and the text, which looks better I think.