Improve translateURL function
The URL is now built in each if function rather than set in each one. This means less code is needed as if a condition is not met the already formed URL can be used (always currently).

Because of the URL being built, I can now just check for /blog/ (and /living-page/) in the URL, rather than blog/2, which is more exact. I then check if blog is the last part of the URL, and if so don't check for a translated page (this is just to make the code more readable/faster, it wouldn't find any page anyway given it's the end of the URL.. unless I called one "blog" or "living-page" perhaps.). As this URL is built in each stage now, this means I'm not relying on the general function at the end to just copy the URL (better if/when the translated name is different), and makes it easier to build in blog pages for years later (something I'm planning on doing).

Living-pages (which aren't published yet) now get correctly redirected. Currently to home, but with the commented out code to go to living-pages and translations if available.

I've also used a different variable for when translating article names - post/page, and translatedPost/Page, to be clearer (and in some cases to not overwrite newURL). Year is now lowercase to match other variables.

All if statements have a new line after them now, as do changes between different blocks of code in functions.

Better comments to explain how it works and when something isn't implemented yet.

Semicolons are used everywhere they should be.