Set manifest display property back to minimal-ui
I was previously using display and display_override because iOS didn't support minimal-ui, and I preferred a standalone window to opening in browser. Now iOS 26 ignores these properties all together and gets the user to choose whether they want to install a page as an app or a bookmark, defaulting to app.

In theory keeping the old properties should be fine, and more sensible, but in practice it's not. Chromium has a bug on Android where minimal-ui doesn't work for display_override, so it wasn't working there, and Firefox doesn't support display_override at all (whereas it does support minimal-ui), so also wasn't working there. This fixes both of those.

There's no browsers which support display_override but not minimal-ui, and that seems quite unlikely to happen. Safari on macOS already ignored this like iOS now does. The only issue with this change is it means the PWA will open in browser on older iOS versions (unless they get 3p browsers with PWAs), which is a shame, but on balance I think worth it for my site. Judging by previous iOS releases, marketshare for iOS 26 should be at about 20% now and will be about 75% by the end of the year.