Adapt Roderick's HTML transform to automatically add HTML height and width to images
These get overruled by CSS styling, which now sets all images to "auto", so that if one dimension is set (eg width: 100%) the other still follows with the aspect-ratio getting messed up. What this does mean is that the browser always knows how big an image is, so can allocate it the right amount of space without ever having any layout shift, automatically with no need to manually set aspect-ratio (or width and height) like I used to typically for above-the-fold images. It's good to do for all images though because a user might see an image which isn't loaded below the initial viewport if they're on a slow connection or if they use a link to a specific section (heading/text fragment).

Also add "updates" script to update dependencies